pyproject.toml
[project]
name = "cryptax"
version = "0.1.0"
description = "Generates tax reports from crypto transactions for different exchanges and valid in various countries."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "dali-rp2>=0.6.11",
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "odfpy>=1.4.1",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "pymysql>=1.1.2",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "redis>=7.4.0",
    "rp2>=1.7.2",
    "rq>=2.8.0",
    "sphinx>=9.1.0",
    "sphinx-rtd-theme>=3.1.0",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.45.0",
]
